Moved edit charging switch to config editor
